requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

bcrypt==4.3.0
fastapi==0.115.9
langchain==0.3.25
langchain-core==0.3.60
langchain-google-genai==2.1.4
pydantic==2.11.4
PyJWT==2.10.1
pymongo==4.13.0
python-dotenv==1.1.0
starlette==0.45.3
uvicorn==0.34.2
langchain-community==0.3.24
